+++ docker/mxm-20-data-assetclasses.sql
@@ -1,0 +1,9 @@
+-- TethysDash:
+insert into asset_classes (class_name, description) values ('LRAUV', 'Long-Range Autonomous Underwater Vehicle');
+
+-- TSAUV Front tracking:
+insert into asset_classes (class_name, description) values ('Waveglider', 'LR Wave Glider');
+
+-- mxm-provider-example:
+insert into asset_classes (class_name, description) values ('AcmeDevice', 'Used by mxm-provider-example');
+insert into asset_classes (class_name, description) values ('FooPlat', 'Used by mxm-provider-example');