+++ docker/mxm-25-data-assets.sql
@@ -1,0 +1,12 @@
+-- TSAUV Front tracking:
+insert into assets (asset_id, class_name, description) values ('Tiny', 'Waveglider', 'WG Tiny');
+insert into assets (asset_id, class_name, description) values ('Sparky', 'Waveglider', 'WG Sparky');
+insert into assets (asset_id, class_name, description) values ('SV3-117', 'Waveglider', 'WG SV3-117');
+insert into assets (asset_id, class_name, description) values ('SV3-127', 'Waveglider', 'WG SV3-127');
+
+-- mxm-provider-example:
+insert into assets (asset_id, class_name, description) values ('acme1', 'AcmeDevice', 'Acme device 1');
+insert into assets (asset_id, class_name, description) values ('acme2', 'AcmeDevice', 'Acme device 2');
+insert into assets (asset_id, class_name, description) values ('acme3', 'AcmeDevice', 'Acme device 3');
+insert into assets (asset_id, class_name, description) values ('foo1', 'FooPlat', 'Foo platform 1');
+insert into assets (asset_id, class_name, description) values ('foo2', 'FooPlat', 'Foo platform 2');